Handle 4k pages flag
if the SMALL_PAGES flag is defined it will compile some additional code
in boot.s to map addictional structures needed for 4k pages.
Not all structures are mapped yet.
Not sure if it works yet, but apparently is going in the right
direction, since now the kernel is loaded, and it cause a #PF printing
my #PF handling message.